+++ src/baldur/services/bulkhead/semaphore.py
@@ -36,6 +36,8 @@ from baldur.utils.time import utc_now
 
 logger = structlog.get_logger()
 
+__all__ = ["SemaphoreBulkhead"]
+
 
 class SemaphoreBulkhead(Bulkhead):
     """